refactor(evals): share screenshot hygiene primitives
