+++ evals/packages/cdp/src/index.ts
@@ -1,6 +1,7 @@
 export * from "./app-state.ts";
 export * from "./cdp.ts";
 export * from "./ports.ts";
+export * from "./screen.ts";
 export * from "./screen-state.ts";
 export * from "./surface.ts";
 export * from "./targets.ts";
